Struct inexor_rgf_core_builder::component_builder::ComponentBuilder

pub struct ComponentBuilder { /* private fields */ }

Implementations

impl ComponentBuilder

pub fn new<C: Into<ComponentTypeId>>(ty: C) -> ComponentBuilder

pub fn new_from_type<S: Into<String>>( namespace: S, type_name: S ) -> ComponentBuilder

Creates a builder for creating a new component with the given name and namespace.

pub fn description<S: Into<String>>( &mut self, description: S ) -> &mut ComponentBuilder

Sets the description of the component.

pub fn property<S: Into<String>>( &mut self, property_name: S, data_type: DataType ) -> &mut ComponentBuilder

Adds a property to the component with the given name and data type.

pub fn input_property<S: Into<String>>( &mut self, property_name: S, data_type: DataType ) -> &mut ComponentBuilder

Adds a property to the component with the given name and data type. The socket type is INPUT.

pub fn output_property<S: Into<String>>( &mut self, property_name: S, data_type: DataType ) -> &mut ComponentBuilder

Adds a property to the component with the given name and data type. The socket type is OUTPUT.

pub fn property_from<S: Into<PropertyType>>( &mut self, property_type: S ) -> &mut ComponentBuilder

Adds a property to the component. Uses the given property type.

pub fn string_property<S: Into<String>>( &mut self, property_name: S ) -> &mut ComponentBuilder

Adds a string property to the component with the given name.

pub fn bool_property<S: Into<String>>( &mut self, property_name: S ) -> &mut ComponentBuilder

Adds a boolean property to the component with the given name.

pub fn number_property<S: Into<String>>( &mut self, property_name: S ) -> &mut ComponentBuilder

Adds a number property to the component with the given name.

pub fn array_property<S: Into<String>>( &mut self, property_name: S ) -> &mut ComponentBuilder

Adds an array property to the component with the given name.

pub fn object_property<S: Into<String>>( &mut self, property_name: S ) -> &mut ComponentBuilder

Adds an object property to the component with the given name.

pub fn extension<S: Into<String>>( &mut self, namespace: S, name: S, extension: Value ) -> &mut ComponentBuilder

Adds an extension to the component with the given name.

pub fn build(&self) -> Component

Constructs the component with the previously defined builder data.